@@ -7,12 +7,12 @@
     var pathEnviarMensaje = '@Url.Action("EnviarMensaje")'
 </script>
 <link rel="stylesheet" href="@Url.Content("~/css/contacto.css")"/>
-<div class="mx-5 mt-5">
+<div class="mx-md-5 mt-5">
     <h1 class="text-center font-weight-bold">Contacto</h1>
     <div class="d-md-flex mt-4">
         <div class="col col-md-6 pr-5">
             <div class="mb-5">
-                <p class="text-center font-weight-bold">
+                <p class="text-md-center font-weight-bold">
                     Te brindamos todos los medios necesarios para que puedas comunicarte con nosotros.
                 </p>
             </div>
@@ -30,7 +30,7 @@
                 <div>Concejal Nicolás Defilippi 1253, Ituzaingó, Buenos Aires</div>
             </div>
         </div>
-        <div class="col col-md-6 pl-5">
+        <div class="col col-md-6 pl-md-5">
             <form>
                 <div class="form-group">
                     <label>Nombre</label>
