@@ -4,7 +4,7 @@
     <div class="mx-5 mt-5">
         <h1 class="text-center font-weight-bold">Bombas y motores hidráulicos</h1>
         <div class="d-md-flex mt-4 justify-content-center">
-            <div class="card col mx-3 col-md-5 p-0">
+            <div class="card col mx-md-3 col-md-5 p-0 mb-md-0 mb-3">
                 <img src="@Url.Content("~/img/foto1.jpg")" class="mx-auto" width="75%">
                 <div class="card-body">
                     <h5 class="card-title font-weight-bold text-center">Bombas hidráulicas</h5>
@@ -13,7 +13,7 @@
                     </p>
                 </div>
             </div>
-            <div class="card col mx-3 col-md-5 p-0">
+            <div class="card col mx-md-3 col-md-5 p-0 mb-md-0 mb-3">
                 <img src="@Url.Content("~/img/foto2.jpg")" class="mx-auto" width="60%">
                 <div class="card-body">
                     <h5 class="card-title font-weight-bold text-center">Motores hidráulicos</h5>
