@@ -4,7 +4,7 @@
 <div class="mx-5 mt-5">
     <h1 class="text-center font-weight-bold">Mangueras y accesorios</h1>
     <div class="d-md-flex mt-4 justify-content-center">
-        <div class="card col mx-3 col-md-5 p-0">
+        <div class="card col mx-md-3 col-md-5 p-0 mb-md-0 mb-3">
             <img src="@Url.Content("~/img/c2.jpg")" class="card-img-top" alt="">
             <div class="card-body">
                 <h5 class="card-title text-center font-weight-bold">Mangueras y accesorios</h5>
