@@ -8,7 +8,7 @@
         Válvulas modulares y en línea para todo tipo de industria.
     </h4>
     <div class="d-md-flex mt-4 justify-content-center">
-        <div class="card col mx-3 col-md-5 p-0">
+        <div class="card col mx-md-3 col-md-5 p-0 mb-md-0 mb-3">
             <div class="d-flex justify-content-center modulares">
                 <img src="@Url.Content("~/img/foto4.jpg")" width="50%" class="mx-auto">
             </div>
@@ -17,7 +17,7 @@
             </div>*@
             <h5 class="card-title text-center font-weight-bold py-3">Válvulas modulares</h5>
         </div>
-        <div class="card col mx-3 col-md-5 p-0 hid-mov">
+        <div class="card col mx-md-3 col-md-5 p-0 hid-mov mb-md-0 mb-3">
             <div class="d-flex justify-content-center">
                 <img src="@Url.Content("~/img/foto5.jpg")" width="100%" class="mx-auto">
             </div>
